feat: Initialize Peninsula stack with Docker setup including PostgreSQL, PrestaShop, and API
- Add .env.example for environment variable configuration
- Create .gitignore to exclude sensitive files
- Document deployment process in DEPLOY.md
- Implement Dockerfile and entrypoint for API service
- Set up docker-compose.yml for multi-container orchestration
- Add PrestaShop module for product synchronization with Peninsula API
- Implement health checks and logging for services
- Include backup and restore instructions for databases

+++ api/src/routes/products.js
@@ -102,7 +102,7 @@ router.get("/", requireAuth, async (req, res) => {
   const result = await query(
     `SELECT p.*,
             c.name AS category_name,
-            COALESCE(s.total_stock, 0) AS total_stock,
+            COALESCE(s.total_stock, 0)::int AS total_stock,
             COALESCE(ch.channels, '[]'::jsonb) AS channels
      FROM products p
      LEFT JOIN categories c ON c.id = p.category_id
